Refactor: remove redundant functions, change package
